# Optimize `PrefixString()`