- pkg/identity: Add PrefixString() tests
- pkg/identity: Add benchmark for PrefixString()
- pkg/identity: Optimize PrefixString()

Co-contributed by @schlosna.

---

Main commit pasted here for ease of viewing:

> Remove the use of fmt.Sprintf() which is known to be inefficient and
> simplify the function by reducing allocations and computations
> generating prefix string.
>
> Results:
>
> ```
> $ go test -v ./pkg/identity -run '^$' -count 10 -bench BenchmarkIPIdentityPair_PrefixString -benchmem > old.txt
> $ go test -v ./pkg/identity -run '^$' -count 10 -bench BenchmarkIPIdentityPair_PrefixString -benchmem > new.txt
> $ benchstat old.txt new.txt
> name                                     old time/op    new time/op    delta
> IPIdentityPair_PrefixString/host-16         438ns ± 6%     126ns ± 6%  -71.12%  (p=0.000 n=10+10)
> IPIdentityPair_PrefixString/not_host-16     248ns ± 8%      61ns ± 3%  -75.55%  (p=0.000 n=10+9)
> 
> name                                     old alloc/op   new alloc/op   delta
> IPIdentityPair_PrefixString/host-16         64.0B ± 0%     32.0B ± 0%  -50.00%  (p=0.000 n=10+10)
> IPIdentityPair_PrefixString/not_host-16     48.0B ± 0%     16.0B ± 0%  -66.67%  (p=0.000 n=10+10)
> 
> name                                     old allocs/op  new allocs/op  delta
> IPIdentityPair_PrefixString/host-16          5.00 ± 0%      2.00 ± 0%  -60.00%  (p=0.000 n=10+10)
> IPIdentityPair_PrefixString/not_host-16      3.00 ± 0%      1.00 ± 0%  -66.67%  (p=0.000 n=10+10)
> ```

---

Related: #19571