@christarazi 2023-01-24
Runtime test hit https://github.com/cilium/cilium/issues/23272.