@christarazi 2023-03-07
Marking ready to merge given that CI has passed (ingress tests have been replaced so no need to run them, see https://github.com/cilium/cilium/pull/24025) and we have approving reviews.